feat(quan): 优化关注功能并调整数据类型
- 将关注数和帖子数的数据类型从 Integer 改为 Long
-优化前端关注按钮的显示和交互逻辑
- 后端增加异步更新圈子关注数的功能
- 新增 QuanUserBarFollowsService 接口和实现类，用于获取圈子关注数

+++ my-project-backend/src/main/java/com/example/entity/dto/QuanBars.java
@@ -16,6 +16,6 @@ public class QuanBars extends BaseModel {
     private Integer secondCategory;
     private String description;
     private String avatar;
-    private Integer followerCount;
-    private Integer postCount;
+    private Long followerCount;
+    private Long postCount;
 }
